Rename SynologyDSM modules to synology_dsm
