+++ .travis.yml
@@ -12,7 +12,7 @@ cache:
 install:
   - pip install pylint
 before_script:
-  - pylint SynologyDSM
+  - pylint synology_dsm
   - ./scripts/check_format.sh
 script:
   - python setup.py install
