--- SynologyDSM/__init__.py
@@ -1,2 +1,0 @@
-"""The python-synology library."""
-from .SynologyDSM import SynologyDSM
