+++ synology_dsm/__init__.py
@@ -1,0 +1,2 @@
+"""The python-synology library."""
+from .synology_dsm import SynologyDSM
