+++ synology_dsm/api/dsm/information.py
